requirements.txt
# Core dependencies for the AFL SGM analysis tool. Requires Python ≥ 3.10.

pydantic>=2.0
duckdb>=0.10.1
polars>=0.20.6
PyYAML>=6.0
tenacity>=8.2
httpx>=0.27
tqdm>=4.64
python-dotenv>=1.0
playwright>=1.41
great_expectations>=0.18

# Additional dependencies used in the enhanced pipeline
numpy>=1.25          # correlation and numerical routines
requests>=2.31       # optional HTTP client for weather/API calls

# Optional: you can pin a specific version of python‑lambda‑test if you write
# tests against LambdaTest. It is not required for ingestion.
# lambdatest~=1.0